feat: feat new view system
