+++ client/pubspec.yaml
@@ -38,11 +38,11 @@ dependencies:
   client_common:
     git:
       url: git@github.com:lenra-io/client-common.git
-      ref: v1.0.0-beta.13
+      ref: v1.0.0-beta.16
   lenra_ui_runner:
     git:
       url: git@github.com:lenra-io/lenra_ui_runner.git
-      ref: v1.0.0-beta.90
+      ref: v1.0.0-beta.93
 dev_dependencies:
   flutter_test:
     sdk: flutter
