+++ server/lib/dev_tool/channels/app_socket.ex
@@ -1,0 +1,5 @@
+defmodule DevTool.AppSocket do
+  use ApplicationRunner.AppSocket,
+    adapter: DevTool.AppAdapter,
+    route_channel: DevTool.RouteChannel
+end
