+++ server/lib/dev_tool/channels/route_channel.ex
@@ -1,6 +1,6 @@
-defmodule DevTool.AppChannel do
+defmodule DevTool.RouteChannel do
   @moduledoc """
     `LenraWeb.AppChannel` handle the app channel to run app and listeners and push to the user the resulted UI or Patch
   """
-  use ApplicationRunner.AppChannel, adapter: DevTool.AppAdapter
+  use ApplicationRunner.RouteChannel
 end
