+++ server/mix.exs
@@ -60,13 +60,14 @@ defmodule DevTool.MixProject do
       {:telemetry_poller, "~> 0.4"},
       {:jason, "~> 1.0"},
       {:plug_cowboy, "~> 2.0"},
-      {:finch, "~> 0.12.0"},
+      {:finch, "~> 0.13.0"},
       {:ecto_sql, "~> 3.4"},
       {:postgrex, "~> 0.15.8"},
-      {:application_runner,
-       git: "https://github.com/lenra-io/application-runner.git",
-       tag: "v1.0.0-beta.77",
-       submodules: true},
+      # {:application_runner,
+      #  git: "https://github.com/lenra-io/application-runner.git",
+      #  tag: "v1.0.0-beta.82",
+      #  submodules: true},
+      {:application_runner, path: "../../application-runner"},
       {:lenra_common, git: "https://github.com/lenra-io/lenra-common.git", tag: "v2.4.0"},
       {:distillery, "~> 2.1"}
     ]
